Places a set of dates withing a regular time series

where the periodicity of the time series is expressed as numbers of days, weeks, months quarters, or years.

cut_date(dates, full_seq = dates, factor = FALSE, ...)

Arguments

dates

a set of dates

full_seq

a full sequence of allowable dates as created by 'full_seq_dates()'. Alternatively a vector of dates will some regular periodicity, that will be used as an input for 'full_seq_dates()', if missing this will be derived from the data itself.

factor

return the result as an ordered factor with the date ranges as a label. if false this returns a date vector where the date is

...

if full_seq is not give, or a plain vector of dates, other options for 'full_seq_dates()' can be set here. E.g. ('fmt="%d/%m/%Y", period="1 week")

Value

a set of dates, representing the start (or end) of the period the date falls into, where the period is defined by 'full_seq' - which is usually defined by 'full_seq_dates()'

Examples

dates = as.Date(c("2020-01-01","2020-02-01","2020-01-15","2020-02-03",NA))
fs = full_seq_dates(dates, "2 days")
dates - cut_date(dates, fs)
#> Time differences in days
#> [1]  0  1  0 NA NA
cut_date(dates,fs,TRUE)
#> [1] 01 Jan — 02 Jan 31 Jan — 01 Feb 15 Jan — 16 Jan <NA>           
#> [5] <NA>           
#> 16 Levels: 01 Jan — 02 Jan < 03 Jan — 04 Jan < ... < 31 Jan — 01 Feb

# A weekly set of dates:
dates2 = Sys.Date() + floor(runif(50,max=10))*7

# in this specific situation the final date is not truncated because the
# input data is seen as an exact match for the whole output period.
cut_date(dates2, fmt = "%d/%b", factor = TRUE)
#>  [1] 18/Jan — 24/Jan 21/Dec — 27/Dec 07/Dec — 13/Dec 30/Nov — 06/Dec
#>  [5] 11/Jan — 17/Jan 30/Nov — 06/Dec 28/Dec — 03/Jan 23/Nov — 29/Nov
#>  [9] 25/Jan — 31/Jan 04/Jan — 10/Jan 18/Jan — 24/Jan 28/Dec — 03/Jan
#> [13] 21/Dec — 27/Dec 14/Dec — 20/Dec 21/Dec — 27/Dec 28/Dec — 03/Jan
#> [17] 28/Dec — 03/Jan 23/Nov — 29/Nov 14/Dec — 20/Dec 14/Dec — 20/Dec
#> [21] 28/Dec — 03/Jan 21/Dec — 27/Dec 28/Dec — 03/Jan 14/Dec — 20/Dec
#> [25] 23/Nov — 29/Nov 14/Dec — 20/Dec 18/Jan — 24/Jan 14/Dec — 20/Dec
#> [29] 30/Nov — 06/Dec 21/Dec — 27/Dec 04/Jan — 10/Jan 04/Jan — 10/Jan
#> [33] 11/Jan — 17/Jan 23/Nov — 29/Nov 18/Jan — 24/Jan 30/Nov — 06/Dec
#> [37] 07/Dec — 13/Dec 07/Dec — 13/Dec 23/Nov — 29/Nov 14/Dec — 20/Dec
#> [41] 18/Jan — 24/Jan 04/Jan — 10/Jan 23/Nov — 29/Nov 28/Dec — 03/Jan
#> [45] 28/Dec — 03/Jan 30/Nov — 06/Dec 14/Dec — 20/Dec 25/Jan — 31/Jan
#> [49] 28/Dec — 03/Jan 21/Dec — 27/Dec
#> 10 Levels: 23/Nov — 29/Nov < 30/Nov — 06/Dec < ... < 25/Jan — 31/Jan

# if the input dates don't line up with the output dates
# there may be incomplete coverage of the first and last category.
# where the cutting results in short periods. In this
# instance the first and last periods are truncated to prevent them
# being counted as complete when they are in fact potentially missing a few days worth of data:
cut_date(dates2, fmt = "%d/%b", factor = TRUE, period = "-2 weeks", anchor="sun")
#>  [1] <NA>            19/Dec — 01/Jan 05/Dec — 18/Dec <NA>           
#>  [5] 02/Jan — 15/Jan <NA>            19/Dec — 01/Jan <NA>           
#>  [9] <NA>            02/Jan — 15/Jan <NA>            19/Dec — 01/Jan
#> [13] 19/Dec — 01/Jan 05/Dec — 18/Dec 19/Dec — 01/Jan 19/Dec — 01/Jan
#> [17] 19/Dec — 01/Jan <NA>            05/Dec — 18/Dec 05/Dec — 18/Dec
#> [21] 19/Dec — 01/Jan 19/Dec — 01/Jan 19/Dec — 01/Jan 05/Dec — 18/Dec
#> [25] <NA>            05/Dec — 18/Dec <NA>            05/Dec — 18/Dec
#> [29] <NA>            19/Dec — 01/Jan 02/Jan — 15/Jan 02/Jan — 15/Jan
#> [33] 02/Jan — 15/Jan <NA>            <NA>            <NA>           
#> [37] 05/Dec — 18/Dec 05/Dec — 18/Dec <NA>            05/Dec — 18/Dec
#> [41] <NA>            02/Jan — 15/Jan <NA>            19/Dec — 01/Jan
#> [45] 19/Dec — 01/Jan <NA>            05/Dec — 18/Dec <NA>           
#> [49] 19/Dec — 01/Jan 19/Dec — 01/Jan
#> Levels: 05/Dec — 18/Dec < 19/Dec — 01/Jan < 02/Jan — 15/Jan