Format a value omitting NA values

Format a value omitting NA values

Usage

.if_present(x, ...)

Arguments

x

a vector of values

...

Named arguments passed on to base::format

trim

logical; if FALSE, logical, numeric and complex values are right-justified to a common width: if TRUE the leading blanks for justification are suppressed.

digits

a positive integer indicating how many significant digits are to be used for numeric and complex x. The default, NULL, uses getOption("digits"). This is a suggestion: enough decimal places will be used so that the smallest (in magnitude) number has this many significant digits, and also to satisfy nsmall. (For more, notably the interpretation for complex numbers see signif.)

nsmall

the minimum number of digits to the right of the decimal point in formatting real/complex numbers in non-scientific formats. Allowed values are 0 <= nsmall <= 20.

justify

should a character vector be left-justified (the default), right-justified, centred or left alone. Can be abbreviated.

cut.names

logical passed to the list method of as.data.frame().

width

default method: the minimum field width or NULL or 0 for no restriction.

AsIs method: the maximum field width for non-character objects. NULL corresponds to the default 12.

na.encode

logical: should NA strings be encoded? Note this only applies to elements of character vectors, not to numerical, complex nor logical NAs, which are always encoded as "NA".

scientific

either a logical specifying whether elements of a real or complex vector should be encoded in scientific format, or an integer penalty (see options("scipen")). Missing values correspond to the current default penalty.

...

further arguments passed to or from other methods.

big.mark,big.interval,small.mark,small.interval,decimal.mark,zero.print,drop0trailing

used for prettying (longish) numerical and complex sequences. Passed to prettyNum: that help page explains the details.

Value

a vector of character values

Examples

try({
.if_present( c(1,2,NA,4,5)/3, digits=2 )
})
#> Error in .if_present(c(1, 2, NA, 4, 5)/3, digits = 2) : 
#>   could not find function ".if_present"