A simple table as a ggplot patchwork object, no customisation allowed

A simple table as a ggplot patchwork object, no customisation allowed

Usage

gg_simple_table(df, pts = 8, font = "sans", unwrapped = FALSE)

Arguments

df

the dataframe with the table data. Column names will become headings

pts

text size in points

font

the font family

unwrapped

- set this to TRUE if you want to add to a patchwork and use patchwork::wrap_plots(p,list(table))

Value

A gtable object (i.e. a grob) optionally wrapped as a patchwork plot.

Examples

if (FALSE) {
  gg_simple_table(tibble::tibble(x=c(1,2,3),y=c(5,4,3)),pts=10)
}