Get unstable version of package

reboot a library reloading from a local development copy if it exists locally alternatively get the most up to date github package.

unstable(pkg = "ggrrr", org = "terminological")

Arguments

pkg

the package to load the unstable version of

org

the github organisation

Value

nothing

Examples

# \donttest{
library(tidyverse)
ggrrr::unstable()
#> ℹ Loading ggrrr
#> ── R CMD build ─────────────────────────────────────────────────────────────────
#> * checking for file ‘/tmp/RtmpBJ5tjH/file349a310c1f98/ggrrr/DESCRIPTION’ ... OK
#> * preparing ‘ggrrr’:
#> * checking DESCRIPTION meta-information ... OK
#> * checking for LF line-endings in source and make files and shell scripts
#> * checking for empty or unneeded directories
#> Removed empty directory ‘ggrrr/output’
#> Omitted ‘LazyData’ from DESCRIPTION
#> * building ‘ggrrr_0.0.0.9020.tar.gz’
#> 
#> Installing package into ‘/tmp/RtmpimB1Q1/temp_libpath28f859770d86’
#> (as ‘lib’ is unspecified)
# }