Places a set of dates withing a regular time series

where the periodicity of the time series is expressed as numbers of days, weeks, months quarters, or years.

cut_date(dates, full_seq = dates, factor = FALSE, ...)

Arguments

dates
a set of dates

full_seq
a full sequence of allowable dates as created by 'full_seq_dates()'. Alternatively a vector of dates will some regular periodicity, that will be used as an input for 'full_seq_dates()', if missing this will be derived from the data itself.

factor
return the result as an ordered factor with the date ranges as a label. if false this returns a date vector where the date is

...
if full_seq is not give, or a plain vector of dates, other options for 'full_seq_dates()' can be set here. E.g. ('fmt="%d/%m/%Y", period="1 week")

Value

a set of dates, representing the start (or end) of the period the date falls into, where the period is defined by 'full_seq' - which is usually defined by 'full_seq_dates()'

Examples

dates = as.Date(c("2020-01-01","2020-02-01","2020-01-15","2020-02-03",NA))
fs = full_seq_dates(dates, "2 days")
dates - cut_date(dates, fs)

#> Time differences in days
#> [1]  0  1  0 NA NA

cut_date(dates,fs,TRUE)

#> [1] 01 Jan — 02 Jan 31 Jan — 01 Feb 15 Jan — 16 Jan <NA>           
#> [5] <NA>           
#> 16 Levels: 01 Jan — 02 Jan < 03 Jan — 04 Jan < ... < 31 Jan — 01 Feb

# A weekly set of dates:
dates2 = Sys.Date() + floor(runif(50,max=10))*7

# in this specific situation the final date is not truncated because the
# input data is seen as an exact match for the whole output period.
cut_date(dates2, fmt = "%d/%b", factor = TRUE)

#>  [1] 13/Sep — 19/Sep 08/Nov — 14/Nov 25/Oct — 31/Oct 20/Sep — 26/Sep
#>  [5] 13/Sep — 19/Sep 11/Oct — 17/Oct 11/Oct — 17/Oct 27/Sep — 03/Oct
#>  [9] 01/Nov — 07/Nov 01/Nov — 07/Nov 08/Nov — 14/Nov 20/Sep — 26/Sep
#> [13] 13/Sep — 19/Sep 04/Oct — 10/Oct 11/Oct — 17/Oct 20/Sep — 26/Sep
#> [17] 11/Oct — 17/Oct 13/Sep — 19/Sep 04/Oct — 10/Oct 15/Nov — 21/Nov
#> [21] 27/Sep — 03/Oct 25/Oct — 31/Oct 01/Nov — 07/Nov 20/Sep — 26/Sep
#> [25] 15/Nov — 21/Nov 01/Nov — 07/Nov 13/Sep — 19/Sep 18/Oct — 24/Oct
#> [29] 25/Oct — 31/Oct 25/Oct — 31/Oct 13/Sep — 19/Sep 27/Sep — 03/Oct
#> [33] 04/Oct — 10/Oct 25/Oct — 31/Oct 11/Oct — 17/Oct 11/Oct — 17/Oct
#> [37] 01/Nov — 07/Nov 15/Nov — 21/Nov 20/Sep — 26/Sep 27/Sep — 03/Oct
#> [41] 25/Oct — 31/Oct 11/Oct — 17/Oct 25/Oct — 31/Oct 25/Oct — 31/Oct
#> [45] 13/Sep — 19/Sep 01/Nov — 07/Nov 01/Nov — 07/Nov 15/Nov — 21/Nov
#> [49] 15/Nov — 21/Nov 04/Oct — 10/Oct
#> 10 Levels: 13/Sep — 19/Sep < 20/Sep — 26/Sep < ... < 15/Nov — 21/Nov

# if the input dates don't line up with the output dates
# there may be incomplete coverage of the first and last category.
# where the cutting results in short periods. In this
# instance the first and last periods are truncated to prevent them
# being counted as complete when they are in fact potentially missing a few days worth of data:
cut_date(dates2, fmt = "%d/%b", factor = TRUE, period = "-2 weeks", anchor="sun")

#>  [1] <NA>            <NA>            24/Oct — 06/Nov <NA>           
#>  [5] <NA>            10/Oct — 23/Oct 10/Oct — 23/Oct 26/Sep — 09/Oct
#>  [9] 24/Oct — 06/Nov 24/Oct — 06/Nov <NA>            <NA>           
#> [13] <NA>            26/Sep — 09/Oct 10/Oct — 23/Oct <NA>           
#> [17] 10/Oct — 23/Oct <NA>            26/Sep — 09/Oct <NA>           
#> [21] 26/Sep — 09/Oct 24/Oct — 06/Nov 24/Oct — 06/Nov <NA>           
#> [25] <NA>            24/Oct — 06/Nov <NA>            10/Oct — 23/Oct
#> [29] 24/Oct — 06/Nov 24/Oct — 06/Nov <NA>            26/Sep — 09/Oct
#> [33] 26/Sep — 09/Oct 24/Oct — 06/Nov 10/Oct — 23/Oct 10/Oct — 23/Oct
#> [37] 24/Oct — 06/Nov <NA>            <NA>            26/Sep — 09/Oct
#> [41] 24/Oct — 06/Nov 10/Oct — 23/Oct 24/Oct — 06/Nov 24/Oct — 06/Nov
#> [45] <NA>            24/Oct — 06/Nov 24/Oct — 06/Nov <NA>           
#> [49] <NA>            26/Sep — 09/Oct
#> Levels: 26/Sep — 09/Oct < 10/Oct — 23/Oct < 24/Oct — 06/Nov