Expand a data vector to the full range

Convert a vector of observation dates to a ordered sequence of every day in the time series

full_seq_dates(dates, period = "1 day", anchor = "start", fmt = "%d %b")

Arguments

dates
a vector of dates, possibly including NA values

period
the gap between observations as a number or, a negative number means the resulting sequence defines a end of time periods, a positive defines the beginning. may be an integer number of days, or a text string like '2 weeks', '-1 month', etc.

anchor
defines the day of week the periods start or end. either "start", "end", a day of the week, or a date

fmt
a strptime formatting string for date range labels.

Value

a vector of dates for complete between the minimum and maximum of dates, on the day of week of the anchoring date

Examples

full_seq_dates(c("2020-01-01","2020-02-01","2020-01-15","2020-02-01",NA), "2 days")

#> 01 Jan — 02 Jan 03 Jan — 04 Jan 05 Jan — 06 Jan 07 Jan — 08 Jan 09 Jan — 10 Jan 
#>    "2020-01-01"    "2020-01-03"    "2020-01-05"    "2020-01-07"    "2020-01-09" 
#> 11 Jan — 12 Jan 13 Jan — 14 Jan 15 Jan — 16 Jan 17 Jan — 18 Jan 19 Jan — 20 Jan 
#>    "2020-01-11"    "2020-01-13"    "2020-01-15"    "2020-01-17"    "2020-01-19" 
#> 21 Jan — 22 Jan 23 Jan — 24 Jan 25 Jan — 26 Jan 27 Jan — 28 Jan 29 Jan — 30 Jan 
#>    "2020-01-21"    "2020-01-23"    "2020-01-25"    "2020-01-27"    "2020-01-29"