Save a table to a variety of formats

depending on the context return the correct format for a document. The basic output here is to use HTML as an output if possible and convert it to an image or a PDF that can then be included into a latex document for example.

hux_save_as(
  hux,
  filename,
  size = std_size$full,
  maxWidth = size$width,
  maxHeight = size$height,
  aspectRatio = maxWidth/maxHeight,
  formats = c("html", "png", "pdf"),
  defaultFontSize = 8,
  sheetname = fs::path_ext_remove(fs::path_file(filename))
)

Arguments

hux
the huxtable to save

filename
the filename, which may omit the extension

size
a ggrrr::std_size

maxWidth
or the maximum width in inches

maxHeight
and either the maximum height in inches

aspectRatio
or the minimum allowed aspect ratio

formats
if the extension is omitted, all the formats described here will be saved. Currently supported outputs are "html","png","pdf","docx","xlsx"

defaultFontSize
the default font size

sheetname
if saving as an xlsx file.

Value

the output depends on if the function is called in a knitr session. It maybe the HTML or a link to the pdf output for example.